refactor(assumptions): migrate order handlers to new design
Migrate Q.nonnegative and Q.nonzero
