+++ sympy/assumptions/ask.py
@@ -172,46 +172,8 @@ class AssumptionKeys:
 
     @memoize_property
     def nonzero(self):
-        """
-        Nonzero real number predicate.
-
-        Explanation
-        ===========
-
-        ``ask(Q.nonzero(x))`` is true iff ``x`` is real and ``x`` is not zero.  Note in
-        particular that ``Q.nonzero(x)`` is false if ``x`` is not real.  Use
-        ``~Q.zero(x)`` if you want the negation of being zero without any real
-        assumptions.
-
-        A few important facts about nonzero numbers:
-
-        - ``Q.nonzero`` is logically equivalent to ``Q.positive | Q.negative``.
-
-        - See the documentation of ``Q.real`` for more information about
-          related facts.
-
-        Examples
-        ========
-
-        >>> from sympy import Q, ask, symbols, I, oo
-        >>> x = symbols('x')
-        >>> print(ask(Q.nonzero(x), ~Q.zero(x)))
-        None
-        >>> ask(Q.nonzero(x), Q.positive(x))
-        True
-        >>> ask(Q.nonzero(x), Q.zero(x))
-        False
-        >>> ask(Q.nonzero(0))
-        False
-        >>> ask(Q.nonzero(I))
-        False
-        >>> ask(~Q.zero(I))
-        True
-        >>> ask(Q.nonzero(oo))
-        False
-
-        """
-        return Predicate('nonzero')
+        from .handlers.order import NonZeroPredicate
+        return NonZeroPredicate()
 
     @memoize_property
     def nonpositive(self):
@@ -252,39 +214,8 @@ class AssumptionKeys:
 
     @memoize_property
     def nonnegative(self):
-        """
-        Nonnegative real number predicate.
-
-        Explanation
-        ===========
-
-        ``ask(Q.nonnegative(x))`` is true iff ``x`` belongs to the set of
-        positive numbers including zero.
-
-        - Note that ``Q.nonnegative`` and ``~Q.negative`` are *not* the same
-          thing. ``~Q.negative(x)`` simply means that ``x`` is not negative,
-          whereas ``Q.nonnegative(x)`` means that ``x`` is real and not
-          negative, i.e., ``Q.nonnegative(x)`` is logically equivalent to
-          ``Q.zero(x) | Q.positive(x)``.  So for example, ``~Q.negative(I)`` is
-          true, whereas ``Q.nonnegative(I)`` is false.
-
-        Examples
-        ========
-
-        >>> from sympy import Q, ask, I
-        >>> ask(Q.nonnegative(1))
-        True
-        >>> ask(Q.nonnegative(0))
-        True
-        >>> ask(Q.nonnegative(-1))
-        False
-        >>> ask(Q.nonnegative(I))
-        False
-        >>> ask(Q.nonnegative(-I))
-        False
-
-        """
-        return Predicate('nonnegative')
+        from .handlers.order import NonNegativePredicate
+        return NonNegativePredicate()
 
     @memoize_property
     def even(self):
@@ -1242,9 +1173,7 @@ _handlers = [
     ("commutative",       "AskCommutativeHandler"),
     ("composite",         "ntheory.AskCompositeHandler"),
     ("even",              "ntheory.AskEvenHandler"),
-    ("nonzero",           "order.AskNonZeroHandler"),
     ("nonpositive",       "order.AskNonPositiveHandler"),
-    ("nonnegative",       "order.AskNonNegativeHandler"),
     ("zero",              "order.AskZeroHandler"),
     ("positive",          "order.AskPositiveHandler"),
     ("prime",             "ntheory.AskPrimeHandler"),
