+++ sympy/assumptions/ask_generated.py
@@ -106,6 +106,7 @@ def get_known_facts_cnf():
         Q.finite | ~Q.rational,
         Q.finite | ~Q.transcendental,
         Q.hermitian | ~Q.real,
+        Q.nonzero | ~Q.negative,
         Q.rational | ~Q.integer,
         Q.real | ~Q.irrational,
         Q.real | ~Q.negative,
@@ -113,13 +114,15 @@ def get_known_facts_cnf():
         Q.integer | ~Q.even,
         Q.integer | ~Q.odd,
         Q.integer | ~Q.prime,
+        Q.nonnegative | ~Q.positive,
+        Q.nonnegative | ~Q.zero,
+        Q.nonzero | ~Q.positive,
         ~Q.algebraic | ~Q.transcendental,
         ~Q.antihermitian | ~Q.hermitian,
         ~Q.finite | ~Q.infinite,
         ~Q.imaginary | ~Q.real,
         ~Q.irrational | ~Q.rational,
         ~Q.negative | Q.nonpositive,
-        ~Q.negative | Q.nonzero,
         Q.real | ~Q.positive,
         Q.real | ~Q.zero,
         Q.invertible | Q.singular,
@@ -132,10 +135,7 @@ def get_known_facts_cnf():
         Q.invertible | ~Q.positive_definite,
         Q.invertible | ~Q.unitary,
         Q.lower_triangular | ~Q.diagonal,
-        Q.nonnegative | ~Q.positive,
-        Q.nonnegative | ~Q.zero,
         Q.nonpositive | ~Q.zero,
-        Q.nonzero | ~Q.positive,
         Q.normal | ~Q.diagonal,
         Q.normal | ~Q.unitary,
         Q.positive | ~Q.prime,
@@ -150,18 +150,18 @@ def get_known_facts_cnf():
         Q.triangular | ~Q.upper_triangular,
         Q.unitary | ~Q.orthogonal,
         Q.upper_triangular | ~Q.diagonal,
+        Q.negative | ~Q.nonzero | Q.positive,
         ~Q.composite | ~Q.prime,
         ~Q.even | ~Q.odd,
         ~Q.invertible | ~Q.singular,
         ~Q.positive | ~Q.zero,
-        Q.negative | Q.positive | ~Q.nonzero,
         Q.negative | Q.zero | ~Q.nonpositive,
         ~Q.integer | Q.even | Q.odd,
+        ~Q.nonnegative | Q.positive | Q.zero,
         Q.algebraic | Q.transcendental | ~Q.complex | ~Q.finite,
         Q.irrational | Q.rational | ~Q.finite | ~Q.real,
         ~Q.real | Q.orthogonal | ~Q.unitary,
         Q.lower_triangular | Q.upper_triangular | ~Q.triangular,
-        Q.positive | Q.zero | ~Q.nonnegative,
         Q.negative | ~Q.real | Q.positive | Q.zero,
         Q.diagonal | ~Q.lower_triangular | ~Q.upper_triangular,
         Q.invertible | ~Q.fullrank | ~Q.square
